friend

+++ config/route.php
@@ -6,6 +6,7 @@ use Slim\Views\Twig;
 use Slim\Views\TwigMiddleware;
 
 require_once __DIR__ . '/../src/Controllers/userController.php';
+require_once __DIR__ . '/../src/Controllers/messageController.php';
 
 
 
@@ -21,7 +22,7 @@ $app->get('/', function (Request $request, Response $response) {
     session_start();
     if(isset($_SESSION["userName"])){
         return $view->render($response, 'index.php',['name' => $_SESSION["userName"], "id" => $_SESSION["userId"]]);
-    }else return $view->render($response, 'index.php');
+    }else return $view->render($response, 'signIn.php');
 });
 
 // route User
@@ -29,9 +30,19 @@ $app->get('/', function (Request $request, Response $response) {
 $app->get('/signUp', function ($request, $response) {
     $view = Twig::fromRequest($request);
     session_start();
-    if(isset($_SESSION["messageError"])){
-        return $view->render($response, 'signUp.php',['messageError' => $_SESSION["messageError"]]);
-    }else return $view->render($response, 'signUp.php');
+    if(isset($_SESSION["userName"])){
+        return $view->render($response, 'index.php',['name' => $_SESSION["userName"]]);
+    }return $view->render($response, 'signUp.php',['messageError' => $_SESSION["messageError"]]);
+});
+
+$app->get('/signIn', function ($request, $response) {
+    $view = Twig::fromRequest($request);
+    session_start();
+    if(isset($_SESSION["userName"])){
+        return $view->render($response, 'index.php',['name' => $_SESSION["userName"]]);
+    }else if(isset($_SESSION["messageError"])){
+     return $view->render($response, 'signIn.php',['messageError' => $_SESSION["messageError"]]);
+    }else return $view->render($response, 'signIn.php');
 });
 
 $app->post('/user', function (Request $request, Response $response, array $args) use ($app) {
@@ -41,17 +52,64 @@ $app->post('/user', function (Request $request, Response $response, array $args)
     return $response;
 })->add(redirectMiddleware::class);
 
-$app->get('/deleteUser/{id}', function (Request $request, Response $response, array $args) {
-    $uc = new UserController($this->get(EntityManager::class));
-    $uc->deleteUser($args['id']);
-    return $response;
+$app->get('/messagerie', function (Request $request, Response $response) {
+    $view = Twig::fromRequest($request);
+    session_start();
+    
+    if(isset($_SESSION["userName"])){
+        $mc = new MessageController($this->get(EntityManager::class));
+        $messages = $mc->getAll();
+        return $view->render($response, 'messagerie.php', ['messages' => $messages]);
+    }else return $view->render($response, 'signIn.php', ['messageError' => 'Vous devez être connecté']);
 });
+
+$app->get('/messagerie/new', function (Request $request, Response $response) {
+    $view = Twig::fromRequest($request);
+    session_start();
+    if(isset($_SESSION["userName"])){
+        $mc = new MessageController($this->get(EntityManager::class));
+        $messages = $mc->getAll();
+        return $view->render($response, 'newMessage.php',['messageSuccess' => $_SESSION['messageSuccess'],'messageError' => $_SESSION["messageError"]]);
+    }else return $view->render($response, 'signIn.php', ['messageError' => 'Vous devez être connecté']);
+});
+
+
+$app->post('/authentication', function (Request $request, Response $response, array $args) use ($app) {
+    $uc = new UserController($this->get(EntityManager::class));
+    $parsedBody = $request->getParsedBody();
+    $uc->login($parsedBody);
+    return $response;
+})->add(redirectMiddleware::class);
+
+$app->post('/send', function (Request $request, Response $response, array $args) use ($app) {
+    $mc = new MessageController($this->get(EntityManager::class));
+    $parsedBody = $request->getParsedBody();
+    $mc->createMessage($parsedBody);
+    return $response;
+})->add(redirectMiddleware::class);
+
+$app->get('/deleteUser/{id}', function (Request $request, Response $response, array $args) {
+    session_start();
+    if(isset($_SESSION["userId"])){
+        if($_SESSION["userId"] == $args['id']){
+            $uc = new UserController($this->get(EntityManager::class));
+            $uc->deleteUser($args['id']);
+            return $response;
+        }else{ 
+            $_SESSION["header"] = "Location:http://localhost:8080/friend";
+            return $response;
+        }
+    }else{
+        $_SESSION["header"] = "Location:http://localhost:8080/friend";
+        return $response;
+    }
+})->add(redirectMiddleware::class);
 
 $app->get('/signOut', function (Request $request, Response $response, array $args) {
     $uc = new UserController($this->get(EntityManager::class));
     $uc->signOut();
     return $response;
-});
+})->add(redirectMiddleware::class);
 
 $app->get('/friend', function (Request $request, Response $response) {
     $view = Twig::fromRequest($request);
@@ -60,6 +118,32 @@ $app->get('/friend', function (Request $request, Response $response) {
         $uc = new FriendController($this->get(EntityManager::class));
         $listFriends = $uc->listFriends($_SESSION["userId"]);
         $listPending = $uc->pendingList($_SESSION["userId"]);
-        return $view->render($response, 'friends.php',['listFriends' => $listFriends, 'listPending' => $listPending, 'name' => $_SESSION["userName"], "id" => $_SESSION["userId"]]);
+        $listRequest = $uc->requestList($_SESSION["userId"]);
+        return $view->render($response, 'friends.php',['listFriends' => $listFriends, 'listPending' => $listPending, 'listRequest' => $listRequest, 'name' => $_SESSION["userName"], "id" => $_SESSION["userId"]]);
     }
+});
+
+$app->post('/friend', function (Request $request, Response $response) {
+    $uc = new FriendController($this->get(EntityManager::class));
+    $parsedBody = $request->getParsedBody();
+    $uc->createFriend($parsedBody);
+    return $response;
+})->add(redirectMiddleware::class);
+
+$app->get('/deleteFriend/{id}', function (Request $request, Response $response, array $args) {
+    session_start();
+    if(isset($_SESSION["userId"])){
+            $uc = new FriendController($this->get(EntityManager::class));
+            $uc->deleteFriend($args['id']);
+            return $response;
+    }else{ 
+        $_SESSION["header"] = "Location:http://localhost:8080/friend";
+        return $response;
+    }
+})->add(redirectMiddleware::class);
+
+$app->get('/deleteMessage/{id}', function (Request $request, Response $response, array $args) {
+    $mc = new MessageController($this->get(EntityManager::class));
+    $mc->deleteMessage($args['id']);
+    return $response;
 });
